Reformatted reader && optimised flist build
Removed the stupid way of separately listing
directories for recursive calls, which is now
made in-place by checking the flist just once.

Also reversed the order in which the flist is
constructed, as the OS apparently gives the
file list in more or less lexicographical order.
As far as I can tell, only difference are files
containing CAPS.