Fixed path building
Previously the path was parsed and re-parsed time
and time again, causing overly complicated code and
bugs here and there. Now, a full path is constructed
ASAP when it is possible, and from that path the
basename and dirname are extracted when necessary.
WIP: Needs free() checks and stress testing